fix(benchmarks): test trace helpers without importing the CLI
