+++ benchmarks/router/real_data_priority_benchmark.py
@@ -15,27 +15,15 @@ import subprocess
 
 import matplotlib.pyplot as plt
 import numpy as np
-
-if __package__:
-    from .common import (
-        add_common_args,
-        add_synthesis_args,
-        get_aiperf_cmd_for_trace,
-        prepare_trace_dataset,
-        resolve_tokenizer,
-        set_trace_agent_hint,
-        setup_logger,
-    )
-else:
-    from common import (
-        add_common_args,
-        add_synthesis_args,
-        get_aiperf_cmd_for_trace,
-        prepare_trace_dataset,
-        resolve_tokenizer,
-        set_trace_agent_hint,
-        setup_logger,
-    )
+from common import (
+    add_common_args,
+    add_synthesis_args,
+    get_aiperf_cmd_for_trace,
+    prepare_trace_dataset,
+    resolve_tokenizer,
+    setup_logger,
+    tag_requests_with_priority,
+)
 
 logger = setup_logger(__name__)
 
@@ -83,16 +71,6 @@ def offset_hash_ids(tier_requests):
             r["hash_ids"] = [h + offset for h in r["hash_ids"]]
             shifted[tier].append(r)
     return shifted
-
-
-def tag_requests_with_priority(requests, priority):
-    """Return request copies with extra.nvext.agent_hints.priority merged in."""
-    tagged_requests = []
-    for request in requests:
-        tagged_request = copy.deepcopy(request)
-        set_trace_agent_hint(tagged_request, "priority", priority)
-        tagged_requests.append(tagged_request)
-    return tagged_requests
 
 
 def write_trace_file(requests, path):
